requirements.txt
aniso8601==3.0.2
anyio==4.9.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
Flask-GraphQL==2.0.1
graphql-core==3.2.6
graphql-relay==0.4.5
graphql-server-core==1.2.0
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
promise==2.3
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
requests==2.32.4
Rx==1.6.3
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
strawberry-graphql==0.275.5
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
